+++ src/main/java/com/example/project_management/model/Project.java
@@ -12,13 +12,16 @@ import java.util.List;
 public class Project {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+
     @NotBlank(message = "Назва не може бути порожньою!")
     @Size(min = 2, max = 50, message = "Назва повинна бути від 2 до 50 символів")
     @Column(nullable = false, length = 50)
     private String name;
+
     @Column(nullable = true, length = 1024)
     private String description;
+
     @Column(nullable = false, length = 20)
     @Enumerated(EnumType.STRING)
     private ProjectStatus status;
@@ -29,10 +32,11 @@ public class Project {
     @OneToMany(mappedBy = "project", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Task> tasks = new ArrayList<>();
 
+
     public Project() {
     }
 
-    public Project(int id, String name, String description, ProjectStatus status, LocalDate startDate, LocalDate endDate, List<Task> tasks) {
+    public Project(Long id, String name, String description, ProjectStatus status, LocalDate startDate, LocalDate endDate, List<Task> tasks) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -42,61 +46,18 @@ public class Project {
         this.tasks = tasks;
     }
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public ProjectStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(ProjectStatus status) {
-        this.status = status;
-    }
-
-    public LocalDate getStartDate() {
-        return startDate;
-    }
-
-    public void setStartDate(LocalDate startDate) {
-        this.startDate = startDate;
-    }
-
-    public LocalDate getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(LocalDate endDate) {
-        this.endDate = endDate;
-    }
-
-    public List<Task> getTasks() {
-        return tasks;
-    }
-
-    public void setTasks(List<Task> tasks) {
-        this.tasks = tasks;
-    }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+    public String getDescription() { return description; }
+    public void setDescription(String description) { this.description = description; }
+    public ProjectStatus getStatus() { return status; }
+    public void setStatus(ProjectStatus status) { this.status = status; }
+    public LocalDate getStartDate() { return startDate; }
+    public void setStartDate(LocalDate startDate) { this.startDate = startDate; }
+    public LocalDate getEndDate() { return endDate; }
+    public void setEndDate(LocalDate endDate) { this.endDate = endDate; }
+    public List<Task> getTasks() { return tasks; }
+    public void setTasks(List<Task> tasks) { this.tasks = tasks; }
 }
-
-
